+++ app/controllers/model/PrepareUploadRequestV2.scala
@@ -25,8 +25,8 @@ case class PrepareUploadRequestV2(
   callbackUrl: String,
   successRedirect: Option[String],
   errorRedirect: Option[String],
-  minimumFileSize: Option[Int],
-  maximumFileSize: Option[Int],
+  minimumFileSize: Option[Long],
+  maximumFileSize: Option[Long],
   expectedContentType: Option[String])
     extends PrepareUpload {
 
@@ -43,13 +43,13 @@ case class PrepareUploadRequestV2(
 
 object PrepareUploadRequestV2 {
 
-  def reads(maxFileSize: Int): Reads[PrepareUploadRequestV2] =
+  def reads(maxFileSize: Long): Reads[PrepareUploadRequestV2] =
     ((JsPath \ "callbackUrl").read[String] and
       (JsPath \ "successRedirect").readNullable[String] and
       (JsPath \ "errorRedirect").readNullable[String] and
-      (JsPath \ "minimumFileSize").readNullable[Int](min(0)) and
-      (JsPath \ "maximumFileSize").readNullable[Int](min(0) keepAnd max(maxFileSize)) and
+      (JsPath \ "minimumFileSize").readNullable[Long](min(0)) and
+      (JsPath \ "maximumFileSize").readNullable[Long](min(0L) keepAnd max(maxFileSize)) and
       (JsPath \ "expectedContentType").readNullable[String])(PrepareUploadRequestV2.apply _)
       .filter(JsonValidationError("Maximum file size must be equal or greater than minimum file size"))(request =>
-        request.minimumFileSize.getOrElse(0) <= request.maximumFileSize.getOrElse(maxFileSize))
+        request.minimumFileSize.getOrElse(0L) <= request.maximumFileSize.getOrElse(maxFileSize))
 }
